Extract setting definitions and split IMatchResult into one-interface-per-file
Also extracts resolveWatchedRoomIds and handleMatchingMessage out of the App class into standalone lib/ functions, keeping the App class thin.
