+++ config/settingDefinitions.ts
@@ -1,0 +1,116 @@
+import { ISetting, SettingType } from '@rocket.chat/apps-engine/definition/settings';
+import {
+    SETTING_PHOTODNA_API_KEY,
+    SETTING_QUARANTINE_CHANNEL,
+    SETTING_LIMIT_ANALYSIS_TO_CHANNELS,
+    SETTING_WATCH_DMS,
+    SETTING_ENABLE_AUTOMATED_REPORT,
+    SETTING_NCMEC_ORGNAME,
+    SETTING_NCMEC_REPORTER_NAME,
+    SETTING_NCMEC_REPORTER_EMAIL,
+    SETTING_NCMEC_USER,
+    SETTING_NCMEC_PASSWORD,
+    SETTING_NCMEC_ENABLE_TEST_MODE,
+} from './Settings';
+
+export const settingDefinitions: Array<ISetting> = [
+    {
+        id: SETTING_PHOTODNA_API_KEY,
+        type: SettingType.STRING,
+        packageValue: '',
+        required: true,
+        public: false,
+        i18nLabel: 'CSEM_Api_Key_Label',
+        i18nDescription: 'CSEM_Api_Key_Description',
+    },
+    {
+        id: SETTING_QUARANTINE_CHANNEL,
+        type: SettingType.STRING,
+        packageValue: 'csem-quarantine',
+        required: true,
+        public: false,
+        i18nLabel: 'CSEM_Quarantine_Target_Channel_Label',
+        i18nDescription: 'CSEM_Quarantine_Target_Channel_Description',
+    },
+    {
+        id: SETTING_LIMIT_ANALYSIS_TO_CHANNELS,
+        type: SettingType.STRING,
+        packageValue: '',
+        required: true,
+        public: false,
+        i18nLabel: 'CSEM_Limit_Analysis_To_Channels_Csv_Label',
+        i18nDescription: 'CSEM_Limit_Analysis_To_Channels_Csv_Description',
+    },
+    {
+        id: SETTING_WATCH_DMS,
+        type: SettingType.BOOLEAN,
+        packageValue: false,
+        required: true,
+        public: false,
+        i18nLabel: 'CSEM_Watch_DMs_Label',
+        i18nDescription: 'CSEM_Watch_DMs_Description',
+    },
+    {
+        id: SETTING_ENABLE_AUTOMATED_REPORT,
+        type: SettingType.BOOLEAN,
+        packageValue: false,
+        required: true,
+        public: false,
+        i18nLabel: 'CSEM_Enable_Automated_Report_Label',
+        i18nDescription: 'CSEM_Enable_Automated_Report_Description',
+    },
+    {
+        id: SETTING_NCMEC_ORGNAME,
+        type: SettingType.STRING,
+        packageValue: 'TestOrg',
+        required: true,
+        public: false,
+        i18nLabel: 'CSEM_NCMEC_OrgName_Label',
+        i18nDescription: 'CSEM_NCMEC_OrgName_Description',
+    },
+    {
+        id: SETTING_NCMEC_REPORTER_NAME,
+        type: SettingType.STRING,
+        packageValue: 'ReporterName',
+        required: true,
+        public: false,
+        i18nLabel: 'CSEM_NCMEC_ReporterName_Label',
+        i18nDescription: 'CSEM_NCMEC_ReporterName_Description',
+    },
+    {
+        id: SETTING_NCMEC_REPORTER_EMAIL,
+        type: SettingType.STRING,
+        packageValue: 'test@example.org',
+        required: true,
+        public: false,
+        i18nLabel: 'CSEM_NCMEC_ReporterEmail_Label',
+        i18nDescription: 'CSEM_NCMEC_ReporterEmail_Description',
+    },
+    {
+        id: SETTING_NCMEC_USER,
+        type: SettingType.STRING,
+        packageValue: '',
+        required: true,
+        public: false,
+        i18nLabel: 'CSEM_NCMEC_User_Label',
+        i18nDescription: 'CSEM_NCMEC_User_Description',
+    },
+    {
+        id: SETTING_NCMEC_PASSWORD,
+        type: SettingType.STRING,
+        packageValue: '',
+        required: true,
+        public: false,
+        i18nLabel: 'CSEM_NCMEC_Password_Label',
+        i18nDescription: 'CSEM_NCMEC_Password_Description',
+    },
+    {
+        id: SETTING_NCMEC_ENABLE_TEST_MODE,
+        type: SettingType.BOOLEAN,
+        packageValue: true,
+        required: true,
+        public: false,
+        i18nLabel: 'CSEM_NCMEC_TestMode_Label',
+        i18nDescription: 'CSEM_NCMEC_TestMode_Description',
+    },
+];
