+++ lib/IAdvancedInfo.ts
@@ -1,0 +1,4 @@
+export interface IAdvancedInfo {
+    Key: string;
+    Value: string;
+}
