+++ lib/IMatchDetails.ts
@@ -1,0 +1,5 @@
+import { IMatchFlag } from './IMatchFlag';
+
+export interface IMatchDetails {
+    MatchFlags: Array<IMatchFlag>;
+}
