+++ lib/IMatchFlag.ts
@@ -1,0 +1,7 @@
+import { IAdvancedInfo } from './IAdvancedInfo';
+
+export interface IMatchFlag {
+    AdvancedInfo?: Array<IAdvancedInfo>;
+    Source?: string;
+    Violations?: Array<string>;
+}
