+++ lib/IMatchOperationStatus.ts
@@ -1,0 +1,4 @@
+export interface IMatchOperationStatus {
+    Code: number;
+    Description: string;
+}
