+++ lib/IMatchResult.ts
@@ -1,4 +1,6 @@
 import { IImageData } from './IImageData';
+import { IMatchOperationStatus } from './IMatchOperationStatus';
+import { IMatchDetails } from './IMatchDetails';
 
 export interface IMatchResult {
     Status: IMatchOperationStatus;
@@ -7,24 +9,4 @@ export interface IMatchResult {
     IsMatch?: boolean;
     MatchDetails?: IMatchDetails;
     ImageData?: IImageData;
-}
-
-export interface IMatchOperationStatus {
-    Code: number;
-    Description: string;
-}
-
-export interface IMatchDetails {
-    MatchFlags: Array<IMatchFlag>;
-}
-
-export interface IMatchFlag {
-    AdvancedInfo?: Array<IAdvancedInfo>;
-    Source?: string;
-    Violations?: Array<string>;
-}
-
-export interface IAdvancedInfo {
-    Key: string;
-    Value: string;
 }
